// src/app/models/FiltroHistorial.ts
export class FiltroHistorial{
    constructor(){
    }    
        id: string;
        fechaFin: string;
        fechaInicio: string;
        dniPaciente:string;
        nombrePaciente: string;
        apellidosPaciente: string;
        dniMedico:string;
        nombreMedico: string;
        apellidosMedico: string;
        colegiado:string;
        migranas:string;
        diabetes:string;
        asma:string;
    }